requirements.txt
fastapi==0.115.5
passlib==1.7.4
pydantic==2.9.2
PyJWT==2.9.0
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
